refactor(ui): standardize color-select imports
- Switch from absolute to relative imports in color-select.tsx
- Migrate ColorSelect usage from old single-color to new colors array 
- Extend full size variant with group and cell height styles
